Conjugate gradient method
This is the conjugate gradient method to compute the soution for the linear system Ax=b, A:Rn×n, b in Rn. The program takes matrix A, vector b and initial guess x_0 and returns the solution x, number of iterations num and the gradient of the objective function at the point  x.  

The code has commentary.